Added MacOSX and Possibly Linux Support For ICONS

+++ src/components/IconTextCellRenderer.java
@@ -9,11 +9,15 @@ import java.nio.file.attribute.FileAttribute;
 
 import javax.swing.Icon;
 import javax.swing.ImageIcon;
+import javax.swing.JFileChooser;
 import javax.swing.JTable;
 import javax.swing.filechooser.FileSystemView;
 import javax.swing.table.DefaultTableCellRenderer;
 
 import org.apache.commons.net.ftp.FTPFile;
+
+import view.Constants;
+import view.View;
 
 public class IconTextCellRenderer extends DefaultTableCellRenderer {
 	private static final long serialVersionUID = 1L;
@@ -41,12 +45,22 @@ public class IconTextCellRenderer extends DefaultTableCellRenderer {
 				if(value.getType() == 1){
 					if(value.getSize() == -1) return new ImageIcon(this.getClass().getResource("/back.png"));
 					Path f = Files.createTempDirectory(value.getName(), new FileAttribute<?>[0]);
-					i = FileSystemView.getFileSystemView().getSystemIcon(f.toFile());
-					f.toFile().delete();
-					f = null;
+					if(Constants.REMOTE_OS.contains("win")){
+						JFileChooser fc = new JFileChooser();
+						i = fc.getUI().getFileView(fc).getIcon(f.toFile());
+					}else{
+						i = FileSystemView.getFileSystemView().getSystemIcon(f.toFile());
+						f.toFile().delete();
+						f = null;
+					}
 				}else{
 					File f = File.createTempFile(getName(), getExtension(value.getName()));
-					i =  FileSystemView.getFileSystemView().getSystemIcon(f);
+					if(Constants.REMOTE_OS.contains("win")){
+						JFileChooser fc = new JFileChooser();
+						i = fc.getUI().getFileView(fc).getIcon(f);
+					}else{
+						i =  FileSystemView.getFileSystemView().getSystemIcon(f);
+					}
 					f.delete();
 				}
 				return i;
@@ -55,7 +69,12 @@ public class IconTextCellRenderer extends DefaultTableCellRenderer {
 			File value = ((LocalTableModel)table.getModel()).getData()[row];
 			if(table.getColumnName(column).equalsIgnoreCase("Name")){
 				if(value == null)return new ImageIcon(this.getClass().getResource("/back.png"));
-				return FileSystemView.getFileSystemView().getSystemIcon(value);
+				if(View.isWindows()){
+					return FileSystemView.getFileSystemView().getSystemIcon(value);
+				}else{
+					JFileChooser fc = new JFileChooser();
+					return fc.getUI().getFileView(fc).getIcon(value);
+				}
 			}
 		}
 		return null;
